+++ src/app.tsx
@@ -11,6 +11,8 @@ import { installStarredEffects } from "~/stores/starred";
 import { installSearchEffects } from "~/stores/searches";
 import { installTripEffects } from "~/stores/trips";
 import { installSettingsEffects } from "~/stores/settings";
+import { installAutoSyncEffects } from "~/stores/autoSync";
+import { installSyncEffects } from "~/stores/sync";
 import { installDismissedEffects } from "~/stores/dismissed";
 
 /**
@@ -26,6 +28,8 @@ import { installDismissedEffects } from "~/stores/dismissed";
  */
 export function App() {
   installSettingsEffects();
+  installSyncEffects();
+  installAutoSyncEffects();
   installDismissedEffects();
   installStarredEffects();
   installTripEffects();